Original Ant System NetLogo model 3.x
Downloaded from the NetLogo community models site. This was written in NetLogo 3.something. NetLogo 4 doesn't run on my machines. I was able to open it in NetLogo 5.3 but the conversion did not fully succeed. I will put that version on top of this one. And then the heavily converted by me  version that runs in NetLogo 6.2.0.