Conversion to NetLogo 6.2.0. It works!
On 2021-10-27 I downloaded Ant System.nlogo from the NetLogo community models site. It was written in NetLogo 3.1.4 and would not run in 5 or 6.

Using NetLogo 5.3.1, I made an  incompletely successful conversion of Ant System.nlogo (written in NetLogo 3.1.4) that I downloaded from the community models of NetLogo. This
version replaces that one on GitHub. I have now manually altered very much of the code and gotten it running in NetLogo 6.2.0. That is what this version is. It seems to be correct and now, safely ensconced in GitHub, I will feel free to make further modifications.

This is what the 6.2.0 automated conversion program from NetLogo 5.3.1 was able to do, followed by my extensive handiwork.