Updated analysis file: Electric prices
Hi Steve,

Attached please find the updated spreadsheet for electric prices, including the Zero Cost Collar calculator. I've checked the formulas -- all seems to be working right, and I've updated the Collar calculator to reflect the Zero-Cost Collar strike price for gas using the current cost, and the lifetime average cost (which are about equal).

ALSO NOTE: I've updated the Treasury rate assumption to 2.5% to reflect prevalent 10-year interest rates, and I've added in the Cost of Carry to the assumption (this is the cost we have to pay someone else to hold NG for us for us to purchase in the future). We assume 0.50% for cost of carry.

All monthly average NG prices for electricity producers have been included -- the data is only available through end of December, 2021. However, I've added a 2022 row, and FOR NOW, set the January 2022 proce equal to the December 2021 price. That fact is noted with the Jan 2022 price being in red text as a placeholder.

I'd think Jan 2022 prices should be available any day now -- EIA prices usually come online 2 months after the end of the month for the price. (For example, March prices become available at end of May.)

That means that changes in energy prices due to the Ukraine war have not been incorporated. However, we see a substantial increase in NG prices over the second half of 2021 -- due to inflation, supply shortages and a booming economy. December month-end price is quite high, at $5.15, relative to the 5-year trailing average of $3.64. In addition, market spot prices for natural gas and pil are up about 30% since year-end, so don't be surprised if we find out in a few months that NG prices for Feb and Mar are closer to $6 or $7.

Please let me know if you have any questions or need anything further at this time.

-Tate